handling the middelware of verfication

+++ routes/web.php
@@ -28,7 +28,7 @@ Auth::routes(['verify'=>true]);
 
 // Route::get('/home', 'HomeController@index')->name('home');
 
-Route::get('/', 'Home\HomeController@topRated')->name('home.index')->middleware(['verified']);
+Route::get('/', 'Home\HomeController@topRated')->name('home.index');
 //-----------------------------------------------------------------------
 //test
 Route::get('/tests/{test}', 'Personal\PersonalController@show');
@@ -50,20 +50,20 @@ Route::group(['namespace' => 'Article'], function () {
     // Controllers Within The "App\Http\Controllers\Article" Namespace
     Route::get('/articles', 'ArticleController@index')->name('articles.index');
     Route::get('/doctor_article/{doctor}', 'ArticleController@doctor_article')->name('doctor.article');
-    Route::get('/article/create', 'ArticleController@create')->name('articles.create')->middleware(['role:Admin|Doctor','auth']);
-    Route::post('/articles/store', 'ArticleController@store')->name('articles.store')->middleware(['role:Admin|Doctor','auth']);
-    Route::get('/articles/{article}/edit', 'ArticleController@edit')->name('articles.edit')->middleware(['role:Admin|Doctor','auth']);
-    Route::put('/articles/{article}', 'ArticleController@update')->name('articles.update')->middleware(['role:Admin|Doctor','auth']);
+    Route::get('/article/create', 'ArticleController@create')->name('articles.create')->middleware(['role:Admin|Doctor','auth','verified']);
+    Route::post('/articles/store', 'ArticleController@store')->name('articles.store')->middleware(['role:Admin|Doctor','auth','verified']);
+    Route::get('/articles/{article}/edit', 'ArticleController@edit')->name('articles.edit')->middleware(['role:Admin|Doctor','auth','verified']);
+    Route::put('/articles/{article}', 'ArticleController@update')->name('articles.update')->middleware(['role:Admin|Doctor','auth','verified']);
     Route::get('/articles/{article}', 'ArticleController@show')->name('articles.show');
-    Route::delete('/articles/{id}', 'ArticleController@destroy')->name('articles.destroy')->middleware(['role:Admin|Doctor', 'auth']);
+    Route::delete('/articles/{id}', 'ArticleController@destroy')->name('articles.destroy')->middleware(['role:Admin|Doctor', 'auth','verified']);
     Route::get('/articles/cat/{cat}', 'ArticleController@category')->name('articles.category');
 });
 //-------------------------------------------------------------------------------
 //comments
-Route::post('/comments/store/{article_id}', 'Comment\CommentController@store')->name('comments.store');
-Route::get('/comments/{comment}/edit', 'Comment\CommentController@edit')->name('comments.edit');
-Route::put('/comments/{comment}', 'Comment\CommentController@update');
-Route::delete('/comment/{comment}', 'Comment\CommentController@destroy')->name('comment.destroy');
+Route::post('/comments/store/{article_id}', 'Comment\CommentController@store')->name('comments.store')->middleware(['auth','verified']);
+Route::get('/comments/{comment}/edit', 'Comment\CommentController@edit')->name('comments.edit')->middleware(['auth','verified']);
+Route::put('/comments/{comment}', 'Comment\CommentController@update')->middleware(['verified']);
+Route::delete('/comment/{comment}', 'Comment\CommentController@destroy')->name('comment.destroy')->middleware(['auth','verified']);
 
 //--------------------------------------------------------
 // profile
@@ -71,7 +71,7 @@ Route::get('/profiles', 'Profile\ProfileController@showMyProfile')->name('profil
 Route::get('/profiles/{profile}/edit', 'Profile\ProfileController@edit')->name('profiles.edit')->middleware(['role:Doctor', 'auth']);
 Route::put('/profiles/{profile}', 'Profile\ProfileController@update')->name('profiles.update')->middleware(['role:Doctor', 'auth']);
 Route::get('/profiles/{profile}', 'Profile\ProfileController@showAnotherProfile')->name('profiles.another');
-Route::post('/rate', 'Profile\ProfileController@addRate')->name('profiles.addRate')->middleware(['auth']);
+Route::post('/rate', 'Profile\ProfileController@addRate')->name('profiles.addRate')->middleware(['auth','verified']);
 
 //-----------------------------------
 //complete profile
@@ -83,16 +83,16 @@ Route::group([
     'namespace' => 'Doctor',
 ], function () {
     Route::get('/doctors', 'DoctorController@index');
-    Route::get('/dashboardDoctors', 'DoctorController@dashboardDoctors')->middleware(['role:Admin', 'auth']);
-    Route::get('/dashboardDoctors/{doctor}', 'DoctorController@dashboardDoctorshow')->middleware(['role:Admin', 'auth']);
-    Route::get('/dashboardWaitingDoctors', 'DoctorController@dashboardWaitingDoctors')->middleware(['role:Admin', 'auth']);
-    Route::post('/doctors/{doctor}', 'DoctorController@approve');
-    Route::delete('/doctors/{doctor}', 'DoctorController@delete')->middleware(['role:Admin', 'auth']);
+    Route::get('/dashboardDoctors', 'DoctorController@dashboardDoctors')->middleware(['role:Admin', 'auth','verified']);
+    Route::get('/dashboardDoctors/{doctor}', 'DoctorController@dashboardDoctorshow')->middleware(['role:Admin', 'auth','verified']);
+    Route::get('/dashboardWaitingDoctors', 'DoctorController@dashboardWaitingDoctors')->middleware(['role:Admin', 'auth','verified']);
+    Route::post('/doctors/{doctor}', 'DoctorController@approve')->middleware(['role:Admin', 'auth','verified']);
+    Route::delete('/doctors/{doctor}', 'DoctorController@delete')->middleware(['role:Admin', 'auth','verified']);
 });
 //-----------------------------------------------------------------------------------
 //assistant
 Route::group([
-    'middleware' => ['auth', 'role:Admin|Doctor'],
+    'middleware' => ['auth', 'role:Admin|Doctor','verified'],
     'namespace' => 'Assistant',
     'prefix' => 'assistants'
 ], function () {
@@ -103,20 +103,25 @@ Route::group([
 });
 //----------------------------------------------------------------------------
 //create routes for and reveal time
+Route::group([
+    'middleware' => ['auth', 'role:Admin|Doctor|Assistant','verified'],
+], function () {
 Route::get('/reveals', 'RevealTime\RevealTimeController@index')->name('reveal.index');
 Route::get('/reveals/create', 'RevealTime\RevealTimeController@create')->name('reveal.create');
 Route::post('/reveals/store', 'RevealTime\RevealTimeController@store');
 Route::get('/reveals/{reveal}/edit', 'RevealTime\RevealTimeController@edit')->name('reveals.edit');
 Route::put('/reveals/{reveal}', 'RevealTime\RevealTimeController@update')->name('reveals.update');
 Route::delete('/reveals/{reveal}', 'RevealTime\RevealTimeController@destroy')->name('reveals.delete');
+
+});
 //-------------------------------------------------------------------------------------
 //reservation
-Route::get('/reservations', 'Reservation\ReservationController@index');
-Route::post('reservations/{reveal}/{doctor}', 'Reservation\ReservationController@store');
-Route::delete('reservations/{reveal}', 'Reservation\ReservationController@softDelete');
+Route::get('/reservations', 'Reservation\ReservationController@index')->middleware(['auth','verified']);
+Route::post('reservations/{reveal}/{doctor}', 'Reservation\ReservationController@store')->middleware(['auth','verified']);
+Route::delete('reservations/{reveal}', 'Reservation\ReservationController@softDelete')->middleware(['auth','verified']);
 //----------------------------------------------------------------------------------
 // Dashboard
-Route::get('/dashboard', 'Dashboard\DashboardController@index');
+Route::get('/dashboard', 'Dashboard\DashboardController@index')->middleware(['role:Admin|Doctor|Assistant','auth','verified']);
 
 //trends
 Route::get('/trends', 'Trend\TrendController@index')->middleware(['role:Admin','auth']);
